Merge branch next-zenbox

See merge request cdleonard/qp!6

+++ qp.h
@@ -533,14 +533,14 @@ typedef unsigned long qp_militime_t;
 
 #define QP_DUMP_VAR(var) do { \
         typeof(var) val = (var); \
-        if (__builtin_types_compatible_p(typeof(val), u64)) { \
-            QP_DUMP_VAR_FMT_VAL(var, "%lld", *((u64*)(&(val)))); \
+        if (__builtin_types_compatible_p(typeof(val), uint64_t)) { \
+            QP_DUMP_VAR_FMT_VAL(var, "%lld", *((uint64_t*)(&(val)))); \
         } else if (__builtin_types_compatible_p(typeof(val), bool)) { \
             QP_DUMP_VAR_FMT_VAL(var, "%s", ((bool)(val)) ? "true" : "false"); \
         } else if (__builtin_types_compatible_p(typeof(val), int)) { \
             QP_DUMP_VAR_FMT_VAL(var, "%d", (int)(val)); \
         } else if (__builtin_types_compatible_p(typeof(val), void*)) { \
-            QP_DUMP_VAR_FMT_VAL(var, "%px", (void*)(val)); \
+            QP_DUMP_VAR_FMT_VAL(var, "%px", (void*)(uintptr_t)(val)); \
         } else { \
             QP_PRINT_LOC("WTF is " #var "?" QP_NL); \
         } \
